Added implementation for exponential packet size.
This patch allows using exponentially distributed
sending and replying packet sizes.
Due to limits in maximum and minimum packet size,
values lower than min are set to the min and values higher than max are
set to the max.
Furthermore, headers added in lower levels of the networking stack are
taken into account when generating the exponentially distributed samples
and then they're subtracted.