dw_client: now -L and -S options accept also the nested syntax to specify offset alongsize data size, e.g.: size=[val/distrib],offset=[val/distrib] (addressing #71)
The old --load-offset and --data-offset options are still supported to specify the default offsets to be used for subsequent commands, whenever the offset is not customized with an explicit use of the new nested syntax.

test_loadstore.sh is accordingly updated to test various examples of the new supported syntax.